requirements.txt
APScheduler==3.6.3
attrs==22.1.0
beautifulsoup4==4.11.1
cachetools==4.2.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
distlib==0.3.6
env==0.1.0
exceptiongroup==1.0.4
filelock==3.8.2
Flask==2.2.2
gunicorn==20.1.0
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
platformdirs==2.6.0
pluggy==1.0.0
pyparsing==3.0.9
pyTelegramBotAPI==4.8.0
pytest==7.2.0
python-dotenv==0.21.0
python-telegram-bot==13.15
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
tomli==2.0.1
tornado==6.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.12
virtualenv==20.17.1
Werkzeug==2.2.2
